// trip-talk/src/enums/constants.ts
export enum PLACEHOLDERS {
  WRITER = "작성자 명을 입력해 주세요.",
  PASSWORD = "비밀번호를 입력해 주세요.",
  COMMENT = "댓글을 입력해 주세요.",
  TITLE = "제목을 입력해 주세요.",
  CONTESTS = "내용을 입력해 주세요.",
  ADDRESS = "주소를 입력해 주세요.",
  DETAIL_ADDRESS = "상세주소",
  URL = "링크를 입력해 주세요.",
}

export enum INPUT_CHILDREN {
  WRITER = "작성자",
  PASSWORD = "비밀번호",
  TITLE = "제목",
  ADDRESS = "주소",
  URL = "유튜브 링크",
}

export enum BUTTON {
  COMMENT_SUBMIT = "댓글 등록",
  ZIP_CODE_SEARCH = "우편번호 검색",
  CANCEL = "취소",
  ACTION = "하기",
}

export enum BOARDS_WRITE {
  REQUIRED_FIELDS = "필수입력 사항 입니다",
}

export enum COMMENT_FORM {
  NO_COMMENTS = "등록된 댓글이 없습니다.",
}